# Conflicts:
#	src/main/java/com/cydeo/service/CategoryService.java
#	src/main/java/com/cydeo/service/impl/CategoryServiceImpl.java

+++ src/main/java/com/cydeo/controller/ClientVendorController.java
@@ -3,6 +3,7 @@ package com.cydeo.controller;
 
 import com.cydeo.dto.ClientVendorDto;
 import com.cydeo.enums.ClientVendorType;
+import com.cydeo.repository.InvoiceRepository;
 import com.cydeo.service.ClientVendorService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -18,14 +19,16 @@ import java.util.Arrays;
 public class ClientVendorController {
 
     private final ClientVendorService clientVendorService;
+    private final InvoiceRepository invoiceRepository;
 
-    public ClientVendorController(ClientVendorService clientVendorService) {
+    public ClientVendorController(ClientVendorService clientVendorService, InvoiceRepository invoiceRepository) {
         this.clientVendorService = clientVendorService;
+        this.invoiceRepository = invoiceRepository;
     }
 
     @GetMapping("/list")
     public String getClientVendors(Model model) {
-        model.addAttribute("clientVendors", clientVendorService.listAllClientVendorsByCompany());
+        model.addAttribute("clientVendors", clientVendorService.listAllClientVendorsWithInvoiceStatusByCompany());
         return "clientVendor/clientVendor-list";
     }
 
@@ -77,13 +80,11 @@ public class ClientVendorController {
         return "redirect:/clientVendors/list";
     }
 
-    @DeleteMapping("/delete/{id}")
-    public String deleteClientVendor(@PathVariable Long id, Model model) {
-        try {
+
+    @GetMapping("/delete/{id}")
+    public String deleteClientVendor(@PathVariable Long id) {
+        if (!clientVendorService.isHasInvoices(id)) {
             clientVendorService.deleteClientVendor(id);
-        } catch (IllegalStateException e) {
-            model.addAttribute("error", e.getMessage());
-            return "redirect:/clientVendors/list";
         }
         return "redirect:/clientVendors/list";
     }
